0.1.0, but with the updated README.md
First (but with the updated README.md!!) offical version of the mod. A TON of new stuff added, but i dont know what because i havent touch this mod in a year i think, also this is probably the last time i even touch this mod again since im not into FNF anymore.